pakistan-raid/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
asttokens==3.0.0
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudscraper==1.2.71
colorama==0.4.6
cryptography==44.0.2
customtkinter==5.2.2
darkdetect==0.8.0
discord==2.3.2
discord.py==2.5.2
dotenv==0.9.9
executing==2.2.0
fake-useragent==2.1.0
Faker==37.0.0
frozenlist==1.5.0
icecream==2.1.4
idna==3.10
luhn==0.2.0
multidict==6.2.0
packaging==24.2
phonenumbers==9.0.1
propcache==0.3.0
psutil==7.0.0
pyasynchat==1.0.4
pyasyncore==1.0.4
pycparser==2.22
pycryptodomex==3.22.0
pyftpdlib==2.0.1
Pygments==2.19.1
pyparsing==3.2.1
pystyle==2.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-nmap==0.7.1
python-whois==0.9.5
pytz==2025.2
pyzipper==0.3.6
requests==2.31.0
requests-toolbelt==1.0.0
scapy==2.5.0
setuptools==78.1.0
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
yarl==1.18.3
